intel_extreme: use int cast to please the grumpy oldtimer gcc2

+++ src/add-ons/kernel/drivers/graphics/intel_extreme/bios.cpp
@@ -144,7 +144,7 @@ get_bios(void)
 		kVBIOSAddress, vbios.memory));
 
 	int vbtOffset = vbios.ReadWord(kVbtPointer);
-	if ((vbtOffset + sizeof(vbt_header)) >= kVBIOSSize) {
+	if ((vbtOffset + (int)sizeof(vbt_header)) >= kVBIOSSize) {
 		TRACE((DEVICE_NAME": bad VBT offset : 0x%x\n", vbtOffset));
 		delete_area(vbios.area);
 		return false;
